Warnings cleanup: eslint/sass, dev-vuln audit, PHP 8.5 runtime warnings
A - quick wins: eslint js/app = 0 warnings (quotes/eqeqeq/new-cap/no-empty),
sass deprecation gone, contributing.md Node 24.

B - dev-tooling vulns: drop abandoned gulp-bytediff (local through2 helper),
add postcss/uuid overrides. npm audit = 0 vulns.

C - PHP 8.5 runtime warnings (end-state Z = log-not-fatal): index.php chained
error handler swallows+logs E_WARNING/E_DEPRECATED, chains real errors to F3.
Fixed ~96 app/+in-tree-ESI sites (nullable params, PDO consts, empty()/?? guards,
ESI empty($body->error) sweep) + 21 template warnings via isset() idiom.
Real bugs fixed: SystemThera wh_exits_outware typo, Route array_walk by-ref key,
ConnectionModel endpoint types guard, setup.html repeat-group ternary fatal.
All 8 edited templates compile-validated; only unfixable vendor/ deprecations remain.

+++ app/Model/Universe/SystemModel.php
@@ -184,7 +184,7 @@ class SystemModel extends AbstractUniverseModel {
      * @return double
      */
     public function set_securityStatus($secStatus){
-        $secStatus = (double)$secStatus;
+        $secStatus = (float)$secStatus;
         // round for trueSec
         $positive = ($secStatus > 0);
         $trueSec = round($secStatus, 1, PHP_ROUND_HALF_DOWN);
